requirements.txt
asgiref==3.5.2
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
confection==0.0.4
crispy-bootstrap5==0.7
cymem==2.0.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.2
django-adminlte3==0.1.6
django-ajax-selects==2.2.0
django-autocomplete-light==3.9.4
django-crispy-forms==1.14.0
django-import-export==3.0.2
django-utils-six==2.0
et-xmlfile==1.1.0
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
MarkupPy==1.14
MarkupSafe==2.1.2
mathparse==0.1.2
murmurhash==1.0.9
mysqlclient==2.1.1
numpy==1.24.1
odfpy==1.4.1
openpyxl==3.0.10
packaging==23.0
pathy==0.10.1
Pillow==9.3.0
preshed==3.0.8
pydantic==1.10.4
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
smart-open==6.3.0
spacy==3.5.0
spacy-legacy==3.0.11
spacy-loggers==1.0.4
SQLAlchemy==1.3.24
sqlparse==0.4.3
srsly==2.4.5
tablib==3.3.0
thinc==8.1.7
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
tzdata==2022.5
urllib3==1.26.14
wasabi==1.1.1
xlrd==2.0.1
xlwt==1.3.0
